feat(options): replace option_utils.py with quantlib_utils.py
- Add QuantLib-based implementation for options calculations
- Move OptionPosition class to quantlib_utils.py and rename to OptionContract
- Implement American-style options support using binomial tree model
- Add comprehensive tests for the new implementation
- Delete the old option_utils.py file and its tests
- Update all references to use the new implementation

This implementation provides more accurate pricing and greeks for American-style
options using QuantLib's binomial tree model, while maintaining compatibility
with the existing codebase.
